requirements.txt
selenium~=3.141.0
pytest~=6.1.1
pytest-html~=2.1.1
python-dotenv~=0.14.0